Provide esm output
In this diff I added esm support via dual mode. The idea is keeping
cjs/esm pair with the same file name and with js and mjs extensions.
Bundler may choose which format to use.

main (dist/index) bundler will choose on of these
- dist/index.js
- dist/index.mjs

/emotion reexports from dist/emotion
- dist/emotion/index.js
- dist/emotion/index.mjs

Ref https://medium.com/@giltayar/native-es-modules-in-nodejs-status-and-future-directions-part-i-ee5ea3001f71

Upgraded babel to be able to use js config.